compiled choose functions into generate file
