--- rickshaw/choose_archetypes.py
@@ -1,166 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Modudle to help generate an archtypes list."""
-import random
-import subprocess
-import json
-import shutil
-import os
-
-from rickshaw.lazyasd import lazyobject
-
-
-DEFAULT_SOURCES = {':agents:Source', ':cycamore:Source'}
-DEFAULT_SINKS = {':agents:Sink', ':cycamore:Sink'}
-
-NICHE_ARCHETYPES = {
-    "mine": {":cycamore:Source"}, #
-    "conversion" : {":cycamore:Storage"}, #
-    "enrichment": {":cycamore:Enrichment"},
-    "fuel_fab" : {":cycamore:FuelFab"},
-    "fuel_fab:uo2": {":cycamore:FuelFab"}, #not the correct archetype currently possibly
-    "fuel_fab:triso": {":cycamore:FuelFab"},
-    "fuel_fab:mox": {":cycamore:FuelFab"},
-    "reactor": {":cycamore:Reactor"},
-    "reactor:fr": {":cycamore:Reactor"},
-    "reactor:lwr": {":cycamore:Reactor"},
-    "reactor:hwr": {":cycamore:Reactor"},
-    "reactor:htgr": {":cycamore:Reactor"},
-    "reactor:rbmk": {":cycamore:Reactor"},
-    "reactor:pb": {":cycamore:Reactor"},
-    "storage": {":cycamore:Sink"}, #
-    "storage:wet": {":cycamore:Sink"}, #
-    "storage:dry": {":cycamore:Sink"}, #
-    "storage:interim": {":cycamore:Sink"}, #
-    "separations": {":cycamore:Separations"},
-    "repository": {":cycamore:Sink"} #
-    }
-
-
-ANNOTATIONS = {}
-
-
-@lazyobject
-def CYCLUS_EXECUTABLE():
-    return shutil.which('cyclus')
-
-
-@lazyobject
-def H5LS_EXECUTABLE():
-    return shutil.which('h5ls')
-
-
-@lazyobject
-def H5_LIBPATH():
-    prefix = os.path.dirname(os.path.dirname(H5LS_EXECUTABLE[:]))
-    lib = os.path.join(prefix, 'lib')
-    return lib
-
-
-@lazyobject
-def CYCLUS_LD_LIB_PATH():
-    prefix = os.path.dirname(os.path.dirname(CYCLUS_EXECUTABLE[:]))
-    lib = os.path.join(prefix, 'lib')
-    lib += ':' + H5_LIBPATH[:]
-    ld_lib_path = lib + ':' + os.environ.get('LD_LIBRARY_PATH', '')
-    return ld_lib_path
-
-
-@lazyobject
-def CYCLUS_ENV():
-    env = dict(os.environ)
-    env['LD_LIBRARY_PATH'] = CYCLUS_LD_LIB_PATH[:]
-    return env
-
-
-
-def choose_archetypes(niches):
-    """Determines the correct archetype from cyclus or cycamore based on the niche
-
-    Parameters
-    ----------
-        niches : list
-            List of sequential niches returned from choose_niches.py
-
-    Returns
-    -------
-
-        arches : list
-            List of assigned archetypes. Same list length as niches.
-    """
-    arches = [random.choice(tuple(NICHE_ARCHETYPES[niches[0]] | DEFAULT_SOURCES))]
-    for niche in niches[1:-1]:
-        a = random.choice(tuple(NICHE_ARCHETYPES[niche]))
-        arches.append(a)
-    if len(niches) > 1:
-        #used to be NICHE_ARCHETYPES[niches][-1]
-        a = random.choice(tuple(NICHE_ARCHETYPES[niches[-1]] | DEFAULT_SINKS))
-        arches.append(a)
-    return arches
-
-def archetype_block(arches):
-    unique_arches = sorted(set(arches))
-    block = {"spec" : []}
-    spec_keys = ["path", "lib", "name"]
-    for a in unique_arches:
-        spec = dict(zip(spec_keys, a.split(":")))
-        block["spec"].append(spec)
-    return block
-
-def generate_archetype(arche, in_commod, out_commod):
-    """Pulls in the metadata for each archetype
-
-        Parameters
-        ----------
-            arche : str
-                The name of the archetype that is being generated.
-            in_commod : str
-                The incommodity received by the specific archetype as
-                determined by choose_commodities.py
-            out_commod : str
-                The outcommodity produced by the specific archetype as
-                determined by choose_commodities.py
-
-        Returns
-        -------
-            config : dict
-                The JSON formatted archetype dictionary to be put in the input file
-    """
-    if arche not in ANNOTATIONS:
-        anno = subprocess.check_output([CYCLUS_EXECUTABLE[:], "--agent-annotations", arche],
-                                       env=CYCLUS_ENV)
-        try:
-            anno = json.loads(anno.decode())
-        except json.decoder.JSONDecodeError:
-            raise RuntimeError("JSON could not decode annotation " + anno.decode())
-        ANNOTATIONS[arche] = anno
-    annotations = ANNOTATIONS[arche]
-    vals = {}
-    #dereference aliases
-    for name, var in list(annotations["vars"].items()):
-        if isinstance(var, str):
-            annotations["vars"][name] = annotations["vars"].pop(var)
-    #fill in and randomly generate state variables        
-    for name, var in annotations["vars"].items():
-        uitype = var.get("uitype", None)
-        var_type = var["type"]
-        if uitype == "range":
-            if "nichedomain" in var:
-                rng = var["nichedomain"].get(niche, var["range"])
-            else:
-                rng = var["range"]
-            val = random.uniform(*rng)
-            vals[name] = val
-        elif uitype == "incommodity" or uitype == ["oneormore", "incommodity"]:
-            vals[name] = in_commod
-        elif uitype == "outcommodity" or uitype == ["oneormore", "outcommodity"]:
-            vals[name] = out_commod
-        elif uitype == "commodity" or uitype == ["oneormore", "commodity"]:
-            raise KeyError("Can't generate to commodity please use incommodity "
-                           "or outcommodity")
-        elif var_type == "double" or var_type == "float":
-            vals[name] = var.get("default", 0.0)
-        elif var_type == "int":
-            vals[name] = var.get("default", 0)
-    alias = arche.rpartition(":")[-1]
-    config = {"name": alias, "config": {alias: vals}}
-    return config
